requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-decouple==3.8
setuptools
requests
psycopg2-binary
stripe
gunicorn
whitenoise
dj-database-url
Pillow==11.3.0
scikit-learn==1.7.2
joblib==1.3.2
pandas==2.3.2
